Firebase transmitters added to cloud messaging service
