+++ microservices/cloud-messaging-service/src/main/kotlin/org/radarbase/appserver/microservices/cloud/messaging/service/transmitter/FcmTransmitter.kt
@@ -1,32 +1,63 @@
-package org.radarbase.appserver.microservices.notification.service.transmitter
+/*
+ * Copyright 2025 King's College London
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.radarbase.appserver.microservices.cloud.messaging.service.transmitter
 
 import com.google.firebase.ErrorCode
 import com.google.firebase.messaging.FirebaseMessagingException
 import com.google.firebase.messaging.MessagingErrorCode
 import jakarta.inject.Inject
 import jakarta.ws.rs.core.Response
+import org.radarbase.appserver.microservices.cloud.messaging.config.CloudMessagingServiceConfig
 import org.radarbase.appserver.microservices.contract.calls.UserServiceContract
 import org.radarbase.appserver.microservices.contract.exception.ProxyResponseException
 import org.radarbase.appserver.microservices.contract.utils.Utils.deserializeDtoFromContract
 import org.radarbase.appserver.microservices.core.dto.fcm.FcmUserDto
+import org.radarbase.appserver.microservices.core.entity.DataMessage
 import org.radarbase.appserver.microservices.core.entity.Message
 import org.radarbase.appserver.microservices.core.entity.Notification
 import org.radarbase.appserver.microservices.core.exception.FcmMessageTransmitException
 import org.radarbase.appserver.microservices.core.fcm.downstream.FcmSender
+import org.radarbase.appserver.microservices.core.fcm.model.FcmDataMessage
 import org.radarbase.appserver.microservices.core.fcm.model.FcmNotificationMessage
+import org.radarbase.appserver.microservices.core.service.FcmDataMessageService
 import org.radarbase.appserver.microservices.core.service.FcmNotificationService
-import org.radarbase.appserver.microservices.core.service.transmitter.NotificationTransmitter
+import org.radarbase.appserver.microservices.core.service.UserService
 import org.radarbase.appserver.microservices.core.utils.requireNotNullField
-import org.radarbase.appserver.microservices.notification.config.NotificationServiceConfig
 import org.slf4j.LoggerFactory
 import java.util.Objects
 
-class FcmNotificationTransmitterImpl @Inject constructor(
+class FcmTransmitter @Inject constructor(
     private val fcmSender: FcmSender,
     private val notificationService: FcmNotificationService,
-    config: NotificationServiceConfig,
-) : NotificationTransmitter {
+    private val dataMessageService: FcmDataMessageService,
+    config: CloudMessagingServiceConfig,
+) : DataMessageTransmitter, NotificationTransmitter {
+
     private val userServiceUrl = config.contract.user
+
+    override suspend fun send(dataMessage: DataMessage) {
+        try {
+            fcmSender.send(createMessageFromDataMessage(dataMessage))
+        } catch (exc: FirebaseMessagingException) {
+            handleFcmException(exc, dataMessage)
+        } catch (exc: Exception) {
+            throw FcmMessageTransmitException("Could not transmit a data message through Fcm. ${exc.message}")
+        }
+    }
 
     override suspend fun send(notification: Notification) {
         try {
@@ -75,7 +106,6 @@ class FcmNotificationTransmitterImpl @Inject constructor(
         }
     }
 
-
     private suspend fun handleFCMErrorCode(errorCode: MessagingErrorCode, message: Message) {
         when (errorCode) {
             MessagingErrorCode.INTERNAL, MessagingErrorCode.QUOTA_EXCEEDED, MessagingErrorCode.INVALID_ARGUMENT, MessagingErrorCode.SENDER_ID_MISMATCH, MessagingErrorCode.THIRD_PARTY_AUTH_ERROR -> {}
@@ -90,6 +120,10 @@ class FcmNotificationTransmitterImpl @Inject constructor(
 
                 logger.warn("The Device for user {} was unregistered.", message.subjectId)
                 notificationService.removeNotificationsForUser(
+                    projectId,
+                    subjectId,
+                )
+                dataMessageService.removeDataMessagesForUser(
                     projectId,
                     subjectId,
                 )
@@ -130,6 +164,7 @@ class FcmNotificationTransmitterImpl @Inject constructor(
             notification.fcmTopic,
             requireNotNullField(user, "Notification's User")::fcmToken,
         )
+
         return FcmNotificationMessage().apply {
             this.to = to
             this.condition = notification.fcmCondition
@@ -143,12 +178,35 @@ class FcmNotificationTransmitterImpl @Inject constructor(
         }
     }
 
+    private suspend fun createMessageFromDataMessage(dataMessage: DataMessage): FcmDataMessage {
+        val userId = requireNotNullField(dataMessage.userId, "Notification's UserId")
+        val user = deserializeDtoFromContract<FcmUserDto>(
+            UserServiceContract.getUserUsingId(userId, userServiceUrl),
+        ) {
+            "user_not_found ; User with id $userId not found when creating a notification in transmitter"
+        }
+
+        val to = Objects.requireNonNullElseGet<String>(
+                dataMessage.fcmTopic,
+                requireNotNullField(user, "Data Message's User")::fcmToken,
+            )
+        return FcmDataMessage().apply {
+            this.to = to
+            this.condition = dataMessage.fcmCondition
+            this.priority = dataMessage.priority
+            this.mutableContent = dataMessage.mutableContent
+            this.deliveryReceiptRequested = IS_DELIVERY_RECEIPT_REQUESTED
+            this.messageId = dataMessage.fcmMessageId.toString()
+            this.timeToLive = Objects.requireNonNullElse(dataMessage.ttlSeconds, DEFAULT_TIME_TO_LIVE)
+            this.data = dataMessage.dataMap
+        }
+    }
+
     companion object {
-        private val logger = LoggerFactory.getLogger(FcmNotificationTransmitterImpl::class.java)
+        private val logger = LoggerFactory.getLogger(FcmTransmitter::class.java)
 
         private const val IS_DELIVERY_RECEIPT_REQUESTED: Boolean = true
         private const val DEFAULT_TIME_TO_LIVE: Int = 2419200 // 4 weeks
-
 
         private fun getNotificationMap(notification: Notification): Map<String, Any> {
             val notificationMap: MutableMap<String, Any> = HashMap()
@@ -178,5 +236,4 @@ class FcmNotificationTransmitterImpl @Inject constructor(
             }
         }
     }
-
 }
